+++ Router/_ServerRouting/Interfaces/IModuleLoader.ts
@@ -1,0 +1,28 @@
+/// <amd-module name="Router/_ServerRouting/IModuleLoader" />
+
+import { IPageSourceNotFound } from './IPageSource';
+
+
+/**
+ * В модуле, который строится на странице может быть метод getDataToRender.
+ * Этот метод вернет данные для страницы.
+ */
+export interface IModuleToRender {
+    getDataToRender: (url: string) => Promise<unknown>;
+}
+
+/**
+ * Интерфейс результата микропроцесса загрузки модуля - если модуль существует и его удалось загрузить
+ */
+ export interface IModuleFound {
+    loadStatus: 'success';
+    module: IModuleToRender;
+}
+
+/**
+ * Интерфейс результата микропроцесса загрузки модуля - если модуля НЕ существует и его НЕ удалось загрузить
+ */
+export interface IModuleNotFound {
+    loadStatus: 'not_found';
+    notFound: IPageSourceNotFound;
+}
